Day 8 Pause
Putting a pause on Day 8 as doing it correctly would take a REALLY long time to write, and I wanna work on something else

What do I mean by "correct"?

Well, if you run analysis on the graph traversals for part 2, you can find a property on the data that everyone was given to exploit that makes a possible solution a lot simpler. This was confirmed by going to the subreddit.

To do this "correctly", that means handling all possible corner cases, it takes significantly longer. This has been something that has always kinda irked me: The clash between "I'm here to write software" and "I'm here to get an answer". I'm always leaning toward the former, the people on the leaderboard always lean to the latter.

In this case, I'm just gonna store what I got an am proud of, and move on.